moved resolvers to top level inside component-model-user
