+++ packages/component-model-user/resolvers.js
@@ -1,12 +1,12 @@
-const User = require('./index')
+const User = require('./index').model
 
 const resolvers = {
   Query: {
     async currentUser(_, vars, ctx) {
       if (!ctx.user) return null
       return User.findOrCreate(ctx.user)
-    }
-  }
+    },
+  },
 }
 
 module.exports = resolvers
